Update CVE mitigation
Removes using Posix attributes.  Windows file system providers need FileAttribute<ACLEntry> for permissions attributes while Linux/OSX/BSD would use the Posix attributes.  Because Windows uses a different security model, attempting to have functionality that specifies attributes working for all types of systems would not work.  In addition, there does not appear to be a way that can extract file attributes from a file and allow those attributes to be applied in future file creation operations.

Adds a static reference to a path and if a temporary file is requested, the temporary path will be created on demand once.  I'm not a huge fan of using statics in this fashion but trying to make such a change work in any other fashion would require more invasive changes, and keeping binary compatibility is more important in this case.
Keeps NIO createTempFile call.  While it appears that using NIO for temp files does not set group permissions on the file in OSX (probably Linux as well? haven't checked), the documentation for NIO createTempFile does not assure that all implementations are expected to behave in that manner.  Rather than having to litigate whether or not an undocumented behavior satisfies the mitigation requirement of a CVE, we may as well use a directory upon which permissions are clearly set.
Removes deleteOnExit() from temp files.  the directory will be removed on exit, so registering each temporary file would be redundant.